Fix up slack Private Messages
The bot would not respond to private messages. We can get this working
by munging the message on the way in. For direct messages addressed to
a room that matches the bot name, we reply to the a room matching the
source user's name, and we fix up the message to look like it was
directly targeted at us